Fix types and formatting in Kyber

+++ crypto_kem/kyber1024/avx2/rejsample.h
@@ -1,11 +1,12 @@
 #ifndef REJSAMPLE_H
 #define REJSAMPLE_H
 
+#include <stddef.h>
 #include <stdint.h>
 
-unsigned int PQCLEAN_KYBER1024_AVX2_rej_uniform(int16_t *r,
-        unsigned int len,
+size_t PQCLEAN_KYBER1024_AVX2_rej_uniform(int16_t *r,
+        size_t len,
         const uint8_t *buf,
-        unsigned int buflen);
+        size_t buflen);
 
 #endif
